+++ agent-inject/agent/annotations_test.go
@@ -19,8 +19,8 @@ func TestInitCanSet(t *testing.T) {
 	pod := testPod(annotations)
 
 	agentConfig := AgentConfig{
-		"foobar-image", "http://foobar:8200", "test", "test", true, "100", "1000",
-		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		"foobar-image", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "http://proxy:3128",
 	}
 	err := Init(pod, agentConfig)
 	if err != nil {
@@ -35,6 +35,7 @@ func TestInitCanSet(t *testing.T) {
 		{annotationKey: AnnotationAgentImage, annotationValue: "foobar-image"},
 		{annotationKey: AnnotationAgentRequestNamespace, annotationValue: "test"},
 		{annotationKey: AnnotationAgentRevokeOnShutdown, annotationValue: "true"},
+		{annotationKey: AnnotationProxyAddress, annotationValue: "http://proxy:3128"},
 	}
 
 	for _, tt := range tests {
@@ -55,8 +56,8 @@ func TestInitDefaults(t *testing.T) {
 	pod := testPod(annotations)
 
 	agentConfig := AgentConfig{
-		"", "http://foobar:8200", "test", "test", true, "", "",
-		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "", "",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 	}
 	err := Init(pod, agentConfig)
 	if err != nil {
@@ -89,8 +90,8 @@ func TestInitError(t *testing.T) {
 	pod := testPod(annotations)
 
 	agentConfig := AgentConfig{
-		"image", "", "authPath", "namespace", true, "100", "1000",
-		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		"image", "", DefaultVaultAuthType, "authPath", "namespace", true, "100", "1000",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 	}
 	err := Init(pod, agentConfig)
 	if err == nil {
@@ -153,8 +154,8 @@ func TestSecretAnnotationsWithPreserveCaseSensitivityFlagOff(t *testing.T) {
 		var patches []*jsonpatch.JsonPatchOperation
 
 		agentConfig := AgentConfig{
-			"", "http://foobar:8200", "test", "test", true, "100", "1000",
-			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+			"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 		}
 		err := Init(pod, agentConfig)
 		if err != nil {
@@ -204,8 +205,8 @@ func TestSecretAnnotationsWithPreserveCaseSensitivityFlagOn(t *testing.T) {
 		var patches []*jsonpatch.JsonPatchOperation
 
 		agentConfig := AgentConfig{
-			"", "http://foobar:8200", "test", "test", true, "100", "1000",
-			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+			"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 		}
 		err := Init(pod, agentConfig)
 		if err != nil {
@@ -291,8 +292,8 @@ func TestSecretLocationFileAnnotations(t *testing.T) {
 			var patches []*jsonpatch.JsonPatchOperation
 
 			agentConfig := AgentConfig{
-				"", "http://foobar:8200", "test", "test", true, "100", "1000",
-				DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+				"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+				DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 			}
 			err := Init(pod, agentConfig)
 			if err != nil {
@@ -376,8 +377,8 @@ func TestSecretTemplateAnnotations(t *testing.T) {
 		var patches []*jsonpatch.JsonPatchOperation
 
 		agentConfig := AgentConfig{
-			"", "http://foobar:8200", "test", "test", true, "100", "1000",
-			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+			"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 		}
 		err := Init(pod, agentConfig)
 		if err != nil {
@@ -436,8 +437,8 @@ func TestTemplateShortcuts(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			pod := testPod(tt.annotations)
 			agentConfig := AgentConfig{
-				"", "http://foobar:8200", "test", "test", true, "100", "1000",
-				DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+				"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+				DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 			}
 			err := Init(pod, agentConfig)
 			if err != nil {
@@ -496,8 +497,8 @@ func TestSecretCommandAnnotations(t *testing.T) {
 	for _, tt := range tests {
 		pod := testPod(tt.annotations)
 		agentConfig := AgentConfig{
-			"", "http://foobar:8200", "test", "test", true, "100", "1000",
-			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+			"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 		}
 		err := Init(pod, agentConfig)
 		if err != nil {
@@ -629,8 +630,8 @@ func TestCouldErrorAnnotations(t *testing.T) {
 		var patches []*jsonpatch.JsonPatchOperation
 
 		agentConfig := AgentConfig{
-			"", "http://foobar:8200", "test", "test", true, "100", "1000",
-			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+			"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 		}
 		err := Init(pod, agentConfig)
 		if err != nil {
@@ -650,8 +651,8 @@ func TestInitEmptyPod(t *testing.T) {
 	var pod *corev1.Pod
 
 	agentConfig := AgentConfig{
-		"foobar-image", "http://foobar:8200", "test", "test", true, "100", "1000",
-		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+		"foobar-image", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+		DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 	}
 	err := Init(pod, agentConfig)
 	if err == nil {
@@ -679,8 +680,8 @@ func TestVaultNamespaceAnnotation(t *testing.T) {
 		var patches []*jsonpatch.JsonPatchOperation
 
 		agentConfig := AgentConfig{
-			"foobar-image", "http://foobar:8200", "test", "test", true, "100", "1000",
-			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext,
+			"foobar-image", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
 		}
 		err := Init(pod, agentConfig)
 		if err != nil {
@@ -791,5 +792,67 @@ func Test_runAsSameID(t *testing.T) {
 			require.Equal(t, tt.expectedErr, err != nil)
 			require.Equal(t, tt.expectedUserID, agent.RunAsUser)
 		})
+	}
+}
+
+func TestAuthConfigAnnotations(t *testing.T) {
+	tests := []struct {
+		annotations        map[string]string
+		expectedAuthConfig map[string]interface{}
+	}{
+		{
+			map[string]string{
+				"vault.hashicorp.com/role": "backwardscompat",
+			},
+			map[string]interface{}{
+				"role": "backwardscompat",
+			},
+		},
+		{
+			map[string]string{
+				"vault.hashicorp.com/role":             "backwardscompat",
+				"vault.hashicorp.com/auth-config-role": "lowerprio",
+			},
+			map[string]interface{}{
+				"role": "backwardscompat",
+			},
+		},
+		{
+			map[string]string{
+				"vault.hashicorp.com/auth-config-name":                                "foo",
+				"vault.hashicorp.com/auth-config-ca-cert":                             "bar",
+				"vault.hashicorp.com/auth-config-client_cert":                         "baz",
+				"vault.hashicorp.com/auth-config-credential_poll_interval":            "1",
+				"vault.hashicorp.com/auth-config-remove_secret_id_file_after_reading": "false",
+			},
+			map[string]interface{}{
+				"name":                                "foo",
+				"ca_cert":                             "bar", // param name dashes converted to underscores for ease
+				"client_cert":                         "baz",
+				"credential_poll_interval":            "1",     // string->int conversion left up to consuming app HCL parser
+				"remove_secret_id_file_after_reading": "false", // string->bool, same as above
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		pod := testPod(tt.annotations)
+		var patches []*jsonpatch.JsonPatchOperation
+
+		agentConfig := AgentConfig{
+			"", "http://foobar:8200", DefaultVaultAuthType, "test", "test", true, "100", "1000",
+			DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "",
+		}
+		err := Init(pod, agentConfig)
+		if err != nil {
+			t.Errorf("got error, shouldn't have: %s", err)
+		}
+
+		agent, err := New(pod, patches)
+		if err != nil {
+			t.Errorf("got error, shouldn't have: %s", err)
+		}
+
+		require.Equal(t, agent.Vault.AuthConfig, tt.expectedAuthConfig, "expected AuthConfig %v, got %v", tt.expectedAuthConfig, agent.Vault.AuthConfig)
 	}
 }