+++ agent-inject/agent/container_env.go
@@ -2,7 +2,6 @@ package agent
 
 import (
 	"encoding/base64"
-	"log"
 
 	corev1 "k8s.io/api/core/v1"
 )
@@ -33,6 +32,20 @@ func (a *Agent) ContainerEnvVars(init bool) ([]corev1.EnvVar, error) {
 		})
 	}
 
+	if a.Vault.LogFormat != "" {
+		envs = append(envs, corev1.EnvVar{
+			Name:  "VAULT_LOG_FORMAT",
+			Value: a.Vault.LogFormat,
+		})
+	}
+
+	if a.Vault.ProxyAddress != "" {
+		envs = append(envs, corev1.EnvVar{
+			Name:  "HTTPS_PROXY",
+			Value: a.Vault.ProxyAddress,
+		})
+	}
+
 	if a.ConfigMapName == "" {
 		config, err := a.newConfig(init)
 		if err != nil {
@@ -44,19 +57,17 @@ func (a *Agent) ContainerEnvVars(init bool) ([]corev1.EnvVar, error) {
 			Name:  "VAULT_CONFIG",
 			Value: b64Config,
 		})
+	}
 
-		// Add IRSA AWS Env variables for vault containers
-		if a.Pod != nil {
-			envMap := a.getEnvsFromContainer(a.Pod)
-			if len(envMap) == 0 || len(envMap) >= 2 {
-				for k, v := range envMap {
-					envs = append(envs, corev1.EnvVar{
-						Name:  k,
-						Value: v,
-					})
-				}
-			} else {
-				log.Println("WARN: Could not find 'AWS ROLE/ AWS_WEB_IDENTITY_TOKEN_FILE' env variables")
+	// Add IRSA AWS Env variables for vault containers
+	if a.Vault.AuthType == "aws" {
+		envMap := a.getAwsEnvsFromContainer(a.Pod)
+		if len(envMap) == 0 || len(envMap) >= 2 {
+			for k, v := range envMap {
+				envs = append(envs, corev1.EnvVar{
+					Name:  k,
+					Value: v,
+				})
 			}
 		}
 	}