+++ agent-inject/agent/container_sidecar_test.go
@@ -31,12 +31,15 @@ func TestContainerSidecarVolume(t *testing.T) {
 
 		// Test adding an extra secret from Kube secrets for reference by Agent config
 		fmt.Sprintf("%s", AnnotationAgentExtraSecret): "extrasecret",
+
+		// Test copying volume mounts from an existing container in the Pod to the agent container
+		fmt.Sprintf("%s", AnnotationAgentCopyVolumeMounts): "foobar",
 	}
 
 	pod := testPod(annotations)
 	var patches []*jsonpatch.JsonPatchOperation
 
-	err := Init(pod, AgentConfig{"foobar-image", "http://foobar:1234", "test", "test", true, "1000", "100", DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext})
+	err := Init(pod, AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", true, "1000", "100", DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, ""})
 	if err != nil {
 		t.Errorf("got error, shouldn't have: %s", err)
 	}
@@ -48,8 +51,8 @@ func TestContainerSidecarVolume(t *testing.T) {
 
 	container, err := agent.ContainerSidecar()
 
-	// One token volume mount, one config volume mount and two secrets volume mounts
-	require.Equal(t, 5, len(container.VolumeMounts))
+	// One token volume mount, one config volume mount, two secrets volume mounts, and one mount copied from main container
+	require.Equal(t, 6, len(container.VolumeMounts))
 
 	require.Equal(
 		t,
@@ -78,6 +81,11 @@ func TestContainerSidecarVolume(t *testing.T) {
 				Name:      extraSecretVolumeName,
 				MountPath: extraSecretVolumePath,
 				ReadOnly:  true,
+			},
+			corev1.VolumeMount{
+				Name:      "tobecopied",
+				MountPath: "/etc/somewhereelse",
+				ReadOnly:  false,
 			},
 		},
 		container.VolumeMounts,
@@ -162,7 +170,7 @@ func TestContainerSidecar(t *testing.T) {
 	pod := testPod(annotations)
 	var patches []*jsonpatch.JsonPatchOperation
 
-	err := Init(pod, AgentConfig{"foobar-image", "http://foobar:1234", "test", "test", false, "1000", "100", DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext})
+	err := Init(pod, AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", false, "1000", "100", DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, "http://proxy:3128"})
 	if err != nil {
 		t.Errorf("got error, shouldn't have: %s", err)
 	}
@@ -177,7 +185,7 @@ func TestContainerSidecar(t *testing.T) {
 		t.Errorf("creating container sidecar failed, it shouldn't have: %s", err)
 	}
 
-	expectedEnvs := 2
+	expectedEnvs := 4
 	if len(container.Env) != expectedEnvs {
 		t.Errorf("wrong number of env vars, got %d, should have been %d", len(container.Env), expectedEnvs)
 	}
@@ -190,11 +198,23 @@ func TestContainerSidecar(t *testing.T) {
 		t.Error("env value empty, it shouldn't be")
 	}
 
-	if container.Env[1].Name != "VAULT_CONFIG" {
-		t.Errorf("env name wrong, should have been %s, got %s", "VAULT_CONFIG", container.Env[1].Name)
+	if container.Env[1].Name != "VAULT_LOG_FORMAT" {
+		t.Errorf("env name wrong, should have been %s, got %s", "VAULT_LOG_FORMAT", container.Env[1].Name)
+	}
+
+	if container.Env[2].Name != "HTTPS_PROXY" {
+		t.Errorf("env name wrong, should have been %s, got %s", "HTTPS_PROXY", container.Env[2].Name)
 	}
 
 	if container.Env[1].Value == "" {
+		t.Error("env value empty, it shouldn't be")
+	}
+
+	if container.Env[3].Name != "VAULT_CONFIG" {
+		t.Errorf("env name wrong, should have been %s, got %s", "VAULT_CONFIG", container.Env[3].Name)
+	}
+
+	if container.Env[2].Value == "" {
 		t.Error("env value empty, it shouldn't be")
 	}
 
@@ -263,7 +283,7 @@ func TestContainerSidecarRevokeHook(t *testing.T) {
 			pod := testPod(annotations)
 			var patches []*jsonpatch.JsonPatchOperation
 
-			err := Init(pod, AgentConfig{"foobar-image", "http://foobar:1234", "test", "test", tt.revokeFlag, "1000", "100", DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext})
+			err := Init(pod, AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", tt.revokeFlag, "1000", "100", DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, ""})
 			if err != nil {
 				t.Errorf("got error, shouldn't have: %s", err)
 			}
@@ -312,7 +332,7 @@ func TestContainerSidecarConfigMap(t *testing.T) {
 	pod := testPod(annotations)
 	var patches []*jsonpatch.JsonPatchOperation
 
-	err := Init(pod, AgentConfig{"foobar-image", "http://foobar:1234", "test", "test", true, "1000", "100", DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext})
+	err := Init(pod, AgentConfig{"foobar-image", "http://foobar:1234", DefaultVaultAuthType, "test", "test", true, "1000", "100", DefaultAgentRunAsSameUser, DefaultAgentSetSecurityContext, ""})
 	if err != nil {
 		t.Errorf("got error, shouldn't have: %s", err)
 	}
@@ -327,7 +347,7 @@ func TestContainerSidecarConfigMap(t *testing.T) {
 		t.Errorf("creating container sidecar failed, it shouldn't have: %s", err)
 	}
 
-	expectedEnvs := 1
+	expectedEnvs := 2
 	if len(container.Env) != expectedEnvs {
 		t.Errorf("wrong number of env vars, got %d, should have been %d", len(container.Env), expectedEnvs)
 	}
@@ -910,6 +930,7 @@ func TestContainerSidecarSecurityContext(t *testing.T) {
 				GroupID:            strconv.FormatInt(tt.startup.runAsGroup, 10),
 				SetSecurityContext: tt.startup.setSecurityContext,
 				SameID:             tt.startup.runAsSameUser,
+				ProxyAddress:       "",
 			}
 
 			tt.annotations[AnnotationVaultRole] = "foobar"